Gemini test 2/requirements.txt
streamlit
google-generativeai
python-dotenv 
langchain 
